feat(RVT-3D): adds drivers
