+++ src/modules/f1/components/overlay.tsx
@@ -7,7 +7,7 @@ export default function Overlay() {
     <div className="w-full absolute bottom-32 left-0 z-10 font-orbitron" >
       <Container>
           <h1 className="text-9xl font-bold mb-4 text-left text-white tracking-wider ">F1 Teams</h1>
-          <p className="text-2xl tracking-wider text-gray-400 text-left mb-3">
+          <p className="text-2xl tracking-wider text-gray-200 text-left mb-3">
             Explore the most iconic Formula 1 teams.
           </p>
           <TeamsCards/>
