--- src/modules/f1/f1-car-scene/components/f1Cars.tsx
@@ -1,3 +1,0 @@
-export default function F1Cars ({ shownIndex, models }: { shownIndex: number, models }){
-  return (<div></div>);
-}
