+++ src/shared/utils/f1-teams.ts
@@ -7,6 +7,8 @@ export interface F1Team {
   description: string;
   founded: string;
   championships: number;
+  scale: number;
+  drivers: string[];
 }
 
 export const F1_TEAMS: F1Team[] = [
@@ -18,7 +20,9 @@ export const F1_TEAMS: F1Team[] = [
     slug: 'ferrari',
     description: 'Ferrari is the oldest and most successful team in Formula 1 history, with a rich heritage dating back to 1950. Known for their iconic red cars and passionate fanbase, Ferrari has won more Constructors\' Championships than any other team.',
     founded: '1950',
-    championships: 16
+    championships: 16,
+    scale: 0.9,
+    drivers:['leclerc', 'hamilton']
   },
   {
     id: 'red-bull',
@@ -28,7 +32,9 @@ export const F1_TEAMS: F1Team[] = [
     slug: 'red-bull-racing',
     description: 'Red Bull Racing has dominated Formula 1 in recent years, winning multiple Constructors\' Championships. Known for their innovative approach and aggressive racing style, they have revolutionized the sport with their engineering excellence.',
     founded: '2005',
-    championships: 6
+    championships: 6,
+    scale: 1.15,
+    drivers:['verstappen', 'tsunoda']
   },
   {
     id: 'mercedes',
@@ -38,6 +44,8 @@ export const F1_TEAMS: F1Team[] = [
     slug: 'mercedes',
     description: 'Mercedes has been a dominant force in Formula 1 since their return as a works team in 2010. With their innovative hybrid power units and exceptional team management, they have secured multiple Constructors\' and Drivers\' Championships.',
     founded: '1954',
-    championships: 8
+    championships: 8,
+    scale: 1.5,
+    drivers:['russell', 'antonelli']
   }
 ];
